Break out font dropdown into its own files
