+++ src/components/font-dropdown/font-dropdown.jsx
@@ -1,0 +1,129 @@
+import classNames from 'classnames';
+import PropTypes from 'prop-types';
+import React from 'react';
+
+import Button from '../button/button.jsx';
+import Dropdown from '../dropdown/dropdown.jsx';
+import InputGroup from '../input-group/input-group.jsx';
+import Fonts from '../../lib/fonts';
+import styles from './font-dropdown.css';
+
+const ModeToolsComponent = props => (
+    <Dropdown
+        className={classNames(styles.modUnselect, styles.fontDropdown)}
+        enterExitTransitionDurationMs={60}
+        popoverContent={
+            <InputGroup className={styles.modContextMenu}>
+                <Button
+                    className={classNames(styles.modMenuItem)}
+                    onClick={props.onChoose}
+                    onMouseOver={props.onHoverSansSerif}
+                >
+                    <span className={styles.sansSerif}>
+                        {props.getTranslatedFontName(Fonts.SANS_SERIF)}
+                    </span>
+                </Button>
+                <Button
+                    className={classNames(styles.modMenuItem)}
+                    onClick={props.onChoose}
+                    onMouseOver={props.onHoverSerif}
+                >
+                    <span className={styles.serif}>
+                        {props.getTranslatedFontName(Fonts.SERIF)}
+                    </span>
+                </Button>
+                <Button
+                    className={classNames(styles.modMenuItem)}
+                    onClick={props.onChoose}
+                    onMouseOver={props.onHoverHandwriting}
+                >
+                    <span className={styles.handwriting}>
+                        {props.getTranslatedFontName(Fonts.HANDWRITING)}
+                    </span>
+                </Button>
+                <Button
+                    className={classNames(styles.modMenuItem)}
+                    onClick={props.onChoose}
+                    onMouseOver={props.onHoverMarker}
+                >
+                    <span className={styles.marker}>
+                        {props.getTranslatedFontName(Fonts.MARKER)}
+                    </span>
+                </Button>
+                <Button
+                    className={classNames(styles.modMenuItem)}
+                    onClick={props.onChoose}
+                    onMouseOver={props.onHoverCurly}
+                >
+                    <span className={styles.curly}>
+                        {props.getTranslatedFontName(Fonts.CURLY)}
+                    </span>
+                </Button>
+                <Button
+                    className={classNames(styles.modMenuItem)}
+                    onClick={props.onChoose}
+                    onMouseOver={props.onHoverPixel}
+                >
+                    <span className={styles.pixel}>
+                        {props.getTranslatedFontName(Fonts.PIXEL)}
+                    </span>
+                </Button>
+                <Button
+                    className={classNames(styles.modMenuItem)}
+                    onClick={props.onChoose}
+                    onMouseOver={props.onHoverChinese}
+                >
+                    <span className={styles.chinese}>
+                        {props.getTranslatedFontName(Fonts.CHINESE)}
+                    </span>
+                </Button>
+                <Button
+                    className={classNames(styles.modMenuItem)}
+                    onClick={props.onChoose}
+                    onMouseOver={props.onHoverJapanese}
+                >
+                    <span className={styles.japanese}>
+                        {props.getTranslatedFontName(Fonts.JAPANESE)}
+                    </span>
+                </Button>
+                <Button
+                    className={classNames(styles.modMenuItem)}
+                    onClick={props.onChoose}
+                    onMouseOver={props.onHoverKorean}
+                >
+                    <span className={styles.korean}>
+                        {props.getTranslatedFontName(Fonts.KOREAN)}
+                    </span>
+                </Button>
+            </InputGroup>
+        }
+        ref={props.componentRef}
+        tipSize={.01}
+        onOpen={props.onOpenDropdown}
+        onOuterAction={props.onClickOutsideDropdown}
+    >
+        <span className={props.getFontStyle(props.font)}>
+            {props.getTranslatedFontName(props.font)}
+        </span>
+    </Dropdown>
+);
+
+ModeToolsComponent.propTypes = {
+    componentRef: PropTypes.func.isRequired,
+    font: PropTypes.string,
+    getFontStyle: PropTypes.func.isRequired,
+    getTranslatedFontName: PropTypes.func.isRequired,
+    onChoose: PropTypes.func.isRequired,
+    onClickOutsideDropdown: PropTypes.func,
+    onHoverChinese: PropTypes.func,
+    onHoverCurly: PropTypes.func,
+    onHoverHandwriting: PropTypes.func,
+    onHoverJapanese: PropTypes.func,
+    onHoverKorean: PropTypes.func,
+    onHoverMarker: PropTypes.func,
+    onHoverPixel: PropTypes.func,
+    onHoverSansSerif: PropTypes.func,
+    onHoverSerif: PropTypes.func,
+    onOpenDropdown: PropTypes.func
+};
+export default ModeToolsComponent;
